Changed link to the extension in the feed
When you click on a link to an extension in an RSS reader, it doesn't take you to the extension in the gallery, it immediately initiates a download.  I don't want to download extensions without visiting the extension in the gallery, reading about it, and even possibly visiting its source code repository.